Add scrollbars to CROW1_pairwise
The window is given a width and height so it does not change between
records. The record frame is embedded in a canvas, which itself is
embedded in a Labelframe. Horizontal and vertical scrollbars are
embedded in the Labelframe and have their scrolling functionality linked
to the canvas. This way, rather than wide data expanding the window or
moving off the screen and becoming inaccessible, the window stays the
same width, with any excess data moving beyond the right-hand side of
the window, which can be scrolled to.